More file renames and tweaks
