+++ include/kvs_client.hpp
